refactor(frontend/layout): move Layout into components/
